requirements.txt
aioftp>=0.21
aiohttp>=3.8
asyncpg>=0.27
discord.py>=2.3
matplotlib>=3.5
python-dotenv>=0.21
